# backend/docs/swagger.yaml
definitions:
  domain.ChatMessage:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      role:
        $ref: '#/definitions/domain.ChatRole'
      user_id:
        type: integer
      vehicle_id:
        type: integer
    type: object
  domain.ChatRole:
    enum:
    - user
    - assistant
    - system
    type: string
    x-enum-varnames:
    - ChatRoleUser
    - ChatRoleAssistant
    - ChatRoleSystem
  domain.DashboardEventPreview:
    properties:
      cost:
        type: number
      event_date:
        type: string
      id:
        type: integer
      mileage_km:
        type: integer
      title:
        type: string
      type:
        $ref: '#/definitions/domain.EventType'
    type: object
  domain.DashboardPredictionSummary:
    properties:
      created_at:
        type: string
      model_version:
        type: string
      part_category:
        type: string
      part_name:
        type: string
      probability:
        type: number
      recommendation:
        type: string
      remaining_days:
        type: integer
      remaining_km:
        type: integer
      risk_level:
        $ref: '#/definitions/domain.RiskLevel'
      risk_score:
        type: integer
    type: object
  domain.DashboardVehicleSummary:
    properties:
      brand:
        type: string
      id:
        type: integer
      mileage_km:
        type: integer
      model:
        type: string
      vin:
        type: string
      year:
        type: integer
    type: object
  domain.EventType:
    enum:
    - trip
    - refuel
    - repair
    - inspection
    - accident
    - recall
    - warning
    - maintenance
    - prediction
    - diagnostic
    - part_replacement
    - note
    type: string
    x-enum-varnames:
    - EventTypeTrip
    - EventTypeRefuel
    - EventTypeRepair
    - EventTypeInspection
    - EventTypeAccident
    - EventTypeRecall
    - EventTypeWarning
    - EventTypeMaintenance
    - EventTypePrediction
    - EventTypeDiagnostic
    - EventTypePartReplacement
    - EventTypeNote
  domain.PartCatalogItem:
    properties:
      category:
        type: string
      code:
        type: string
      created_at:
        type: string
      default_lifetime_days:
        type: integer
      default_lifetime_km:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
  domain.Prediction:
    properties:
      created_at:
        type: string
      explanation:
        type: string
      id:
        type: integer
      model_version:
        type: string
      part_category:
        type: string
      part_code:
        type: string
      part_name:
        type: string
      predicted_next_date:
        type: string
      predicted_next_mileage:
        type: integer
      probability:
        type: number
      recommendation:
        type: string
      remaining_days:
        type: integer
      remaining_km:
        type: integer
      risk_level:
        $ref: '#/definitions/domain.RiskLevel'
      risk_score:
        type: integer
      source:
        $ref: '#/definitions/domain.PredictionSource'
      vehicle_id:
        type: integer
    type: object
  domain.PredictionSource:
    enum:
    - rule_based
    - mock
    - ml_service
    type: string
    x-enum-varnames:
    - PredictionSourceRuleBased
    - PredictionSourceMock
    - PredictionSourceMLService
  domain.RiskLevel:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - RiskLevelLow
    - RiskLevelMedium
    - RiskLevelHigh
...
  domain.VehicleDashboard:
    properties:
      all_predictions:
        items:
          $ref: '#/definitions/domain.Prediction'
        type: array
      current_mileage:
        type: integer
      latest_events:
        items:
          $ref: '#/definitions/domain.DashboardEventPreview'
        type: array
      prediction_summary:
        $ref: '#/definitions/domain.DashboardPredictionSummary'
      status:
        type: string
      total_events_count:
        type: integer
      total_fuel_expenses:
        type: number
      total_maintenance_cost:
        type: number
      total_repairs_count:
        type: integer
      vehicle:
        $ref: '#/definitions/domain.DashboardVehicleSummary'
    type: object
...
  domain.VehiclePart:
    properties:
      catalog_code:
        type: string
      category:
        type: string
      created_at:
        type: string
      id:
        type: integer
      installed_at_mileage_km:
        type: integer
      last_service_date:
        type: string
      last_service_mileage_km:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      vehicle_id:
        type: integer
    type: object
  handler.ErrorResponse:
    properties:
      error:
        example: internal server error
        type: string
    type: object
...
  handler.catalogResponse:
    properties:
      parts:
        items:
          $ref: '#/definitions/domain.PartCatalogItem'
        type: array
    type: object
  handler.chatHistoryResponse:
    properties:
      count:
        type: integer
      limit:
        type: integer
      messages:
        items:
          $ref: '#/definitions/domain.ChatMessage'
        type: array
      offset:
        type: integer
      vehicle_id:
        type: integer
    type: object
  handler.chatRequest:
    properties:
      message:
        example: Когда мне нужно менять масло?
        type: string
    required:
    - message
    type: object
  handler.chatResponse:
    properties:
      message:
        $ref: '#/definitions/domain.ChatMessage'
    type: object
  handler.createPartRequest:
    properties:
      catalog_code:
        example: engine_oil
        type: string
      category:
        example: fluids
        type: string
      installed_at_mileage_km:
        example: 40000
        type: integer
      last_service_date:
        example: "2025-01-15T00:00:00Z"
        type: string
      last_service_mileage_km:
        example: 40000
        type: integer
      name:
        example: Engine oil
        type: string
    required:
    - name
    type: object
...
  handler.vehiclePartsResponse:
    properties:
      parts:
        items:
          $ref: '#/definitions/domain.VehiclePart'
        type: array
    type: object
...
info:
  contact: {}
paths:
...
  /api/parts/catalog:
    get:
      description: Returns all available parts from the catalog
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.catalogResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List parts catalog
      tags:
      - parts
...
  /api/vehicles/{id}/chat:
    post:
      consumes:
      - application/json
      description: Sends a message to the AI assistant in the context of a specific
        vehicle. Returns the AI-generated response.
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      - description: Chat message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.chatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.chatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Send a chat message
      tags:
      - chat
  /api/vehicles/{id}/chat/history:
    get:
      description: Returns paginated chat message history for a vehicle in chronological
        order.
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      - description: Limit (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.chatHistoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get chat history
      tags:
      - chat
  /api/vehicles/{id}/dashboard:
    get:
      consumes:
      - application/json
      description: 'Aggregated vehicle data: stats, events, predictions'
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.VehicleDashboard'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get vehicle dashboard
      tags:
      - dashboard
...
  /api/vehicles/{id}/parts:
    get:
      description: Returns all parts installed on a vehicle
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.vehiclePartsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BasicAuth: []
      summary: List vehicle parts
      tags:
      - parts
    post:
      consumes:
      - application/json
      description: Creates a new part record for the specified vehicle
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      - description: Part payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createPartRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.VehiclePart'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Add a part to a vehicle
      tags:
      - parts
  /api/vehicles/{id}/parts/{partId}:
    delete:
      description: Removes a part from the vehicle
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      - description: Part ID
        in: path
        name: partId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Delete a vehicle part
      tags:
      - parts
...
swagger: "2.0"